For smb_request_write_raw() return values of 0 now indicate success.
Fixed decoding of maximum buffer size; this used to come out  as 0, which made smbfs switch back to the default minimum supported (1024 bytes).

Corrected use of the maximum raw write size as the size used by the client to perform raw reads with. The maximum allowed are 65535 bytes for a raw read request. Likewise, the maximum raw write size is 65535 bytes, too.