The raw read/write operations now respect the "maximum raw size" limit set by the server.
Fixed decoding of maximum buffer size; this used to come out as 0, which made smbfs switch back to the default minimum supported (1024 bytes).

Corrected use of the maximum raw write size as the size used by the client to perform raw reads with. The maximum allowed are 65535 bytes for a raw read request. Likewise, the maximum raw write size is 65535 bytes, too.

Simplified the raw read portion of smba_read() by removing the test that avoids dropping into a read loop below. The loop handles the special case well.

Recalculated the overhead for SMB_COM_READ. It's still 52 bytes, but now the documentation states how this number comes about.

Recalculated the overhead for SMB_COM_WRITE. It's 52 bytes and not 56 bytes.