Updated to version 2.14
Documented the directory cache a bit better, including the reason why the cache may have been invalidated while the directory contents were being read.

The directory read operation now verifies that it does not run out of cache entries while it is processing new directory records it received from the server.

The directory read operation now both detects whether or not the last entry has been read, this information is now also provided to the function which retrieved the directory entries. This change is in support of improved directory entry caching.

The directory cache now keeps better track of what's actually in it. If a directory entry could not be extracted and decoded, the number of directory entries read would not match the number of entries available from the cache, for example.

A flag now tracks if the directory cache is valid or not. Previously, checking for a valid cache involved looking at a NULL pointer.

If allocating memory for the directory cache failed partly, then there may have been invalid string pointers in the cache table, leading to a crash when trying to release the cache. Fixed.

Adding another entry to the cache, invalidating the cache or resetting the cache now all use dedicated functions instead of directly manipulating the assorted data structures.